Spawned request() in its own http module.
